% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_version.R
\name{check_version}
\alias{check_version}
\title{Compare package versions}
\usage{
check_version()
}
\description{
Look for packages installed with [devtools::dev_mode()](devtools::dev_mode) and compare them
with packages installed in the system (using `.libPaths`)
}
\details{
If any similar packages are found, a warning is issued.
}
