% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warn_package_version.R
\name{warn_package_version}
\alias{warn_package_version}
\title{Warn Package Version}
\usage{
warn_package_version(level = 2)
}
\arguments{
\item{level}{level of warning. If `level = 0`, only packages with a version lower than the one installed on the system is issued.
If `level = 1`, only packages with a version lower or equal than the one installed on the system is issued
If `level = 2`, all common packages are issued}
}
\description{
This function warn the user if some package have different versions installed on the system and in dev_mode
}
